ccat: converted into mfd driver
In order to prepare ccat for mainline kernel integration it was converted into
a mfd (multi-function device) driver.

This requires two major changes to ec_ccat:
1. ccat got split into multiple modules, for etherlab only the netdev module is
required to be patched. So we can remove the code for the basic module, gpio,
sram and update from the etherlab repository.
2. Etherlab should replace the ccat_netdev module and not the basic ccat module.
So we have to rename ec_ccat into ec_ccat_netdev.

Installing ec_ccat_netdev becomes a little bit more compilcated as from now on,
you need to install the basic ccat driver from upstream and only load the netdev
part from etherlab. We hope when ccat becomes available in mainline Linux this
won't remain a drawback.